% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.tab_tibble}
\alias{knit_print.tab_tibble}
\title{Print a tab_tibble}
\usage{
\method{knit_print}{tab_tibble}(x, ...)
}
\arguments{
\item{x}{The \code{tab_tibble} object to print}

\item{...}{Passed to subsequent methods}
}
\value{
The result of \code{knit_print(x$table, ...)}: by default (see
\code{knit_print.tab_list()}) a \code{knitr::asis_output()} value, so a chunk
printing a \code{tab_tibble} needs no \code{results='asis'}.
}
\description{
Print a tab_tibble
}
\seealso{
Other knitters:
\code{\link[=knit_print.tab_list]{knit_print.tab_list()}}
}
\concept{knitters}
