% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablist_printers.R
\name{print_tablist_pander}
\alias{print_tablist_pander}
\title{Print a tablist using \code{pander::pander()}}
\usage{
print_tablist_pander(x, caption, ...)
}
\arguments{
\item{name}{description}
}
\description{
Print a tablist using \code{pander::pander()}
}
