% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{knit_print}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{knitr}{\code{\link[knitr:knit_print]{knit_print()}}}
}}

