% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{tabtibble_missing_package_error}
\alias{tabtibble_missing_package_error}
\title{Missing backend package condition}
\usage{
tabtibble_missing_package_error(pkg, backend)
}
\arguments{
\item{pkg}{The name of the missing package.}

\item{backend}{The \code{tabtibble.backend} value that needed it.}
}
\description{
Missing backend package condition
}
\keyword{internal}
