% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.tab_list}
\alias{knit_print.tab_list}
\title{Print a tab_list}
\usage{
\method{knit_print}{tab_list}(
  x,
  ...,
  caption,
  print_fun = print_tablist_pander,
  tab_prefix = NULL,
  tab_suffix = "\\n\\n"
)
}
\arguments{
\item{x}{The \code{tab_list} object to print}

\item{...}{passed to \code{print_fun}}

\item{caption}{The caption for each table as a character vector}

\item{print_fun}{A function taking arguments of \code{x} (one data.frame to
print), \code{caption} (the caption for that data.frame), and \code{...}.}

\item{tab_prefix, tab_suffix}{Any text to add before/after each figure (\code{NULL}
to omit)}
}
\value{
\code{x} invisibly
}
\description{
Print a tab_list
}
\seealso{
Other knitters: 
\code{\link{knit_print.tab_tibble}()}
}
\concept{knitters}
