% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print.tab_tibble}
\alias{knit_print.tab_tibble}
\title{Print a tab_tibble}
\usage{
\method{knit_print}{tab_tibble}(x, ...)
}
\arguments{
\item{x}{The \code{tab_tibble} object to print}
}
\value{
\code{x} invisibly
}
\description{
Print a tab_tibble
}
\seealso{
Other knitters: 
\code{\link{knit_print.tab_list}()}
}
\concept{knitters}
