% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{new_tab_tibble}
\alias{new_tab_tibble}
\alias{new_tab_list}
\title{Create a new \code{tab_tibble} object}
\usage{
new_tab_tibble(x)

new_tab_list(x)
}
\arguments{
\item{x}{The object to convert}
}
\value{
An object with the desired class
}
\description{
Create a new \code{tab_tibble} object
}
\section{Functions}{
\itemize{
\item \code{new_tab_list()}: Create a new \code{tab_list} object

}}
