% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablist_printers.R
\name{print_tabtibble}
\alias{print_tabtibble}
\alias{print_tabtibble.default}
\title{Print a single table from a tablist}
\usage{
print_tabtibble(x, caption, ...)

\method{print_tabtibble}{default}(x, caption, ...)
}
\arguments{
\item{x}{A table to print}

\item{caption}{The caption for the table}

\item{...}{Passed to \code{pander::pander}}
}
\value{
The result of \code{pander::pander}
}
\description{
Print a single table from a tablist
}
\section{Methods (by class)}{
\itemize{
\item \code{print_tabtibble(default)}: Print a single table from a tablist using \code{pander::pander()}

}}
